+++ Week2Visual/Week2Visual/Exercise2.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace Week2Visual
+{
+	public partial class Exercise2 : Form
+	{
+		public Exercise2()
+		{
+			InitializeComponent();
+		}
+
+		private void InputData_Load(object sender, EventArgs e)
+		{
+
+		}
+	}
+}