+++ changelog.md
@@ -13,3 +13,7 @@
 - The rest of the exercises for A2 practices
 - Week 2 exercise: Superhero fights
 - Week2Visual for visual projects
+
+## 2024-03-28
+### Added
+- Some visual projects stuff, idk